requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
python-socketio>=5.10.0
opencv-python>=4.8.0
ultralytics>=8.0.196
cvzone>=1.6.1
numpy>=1.26.4
pandas>=2.2.0
sympy>=1.12
torch>=2.4.0
torchvision>=0.19.0
requests>=2.31.0